requirements.txt
Flask==2.2.2
joblib==1.2.0
librosa==0.9.2
mysql_connector_repackaged==0.3.1
noisereduce==2.0.1
numpy==1.21.6
scikit_learn==1.1.2
scipy==1.7.3
sounddevice==0.4.5
soundfile==0.11.0
